requirements.txt
# 图像处理和视频处理
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0

# 网络通信
paho-mqtt>=1.6.0
requests>=2.25.0
pika>=1.2.0
kafka-python>=2.0.0

# 配置和日志
PyYAML>=5.4.0
python-json-logger>=2.0.0

# 工具库
click>=8.0.0
schedule>=1.1.0

# Web应用框架
Flask>=2.0.0
Flask-Login>=0.5.0
Werkzeug>=2.0.0

# 数据库
PyMySQL>=1.0.0
cryptography>=3.4.0

# 图像处理
# 工具库
python-dateutil>=2.8.0

fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 新增依赖项
websockets
onnxruntime
APScheduler 